+++ plugin/attribute/attribute.go
@@ -80,6 +80,8 @@ const (
 	FlyingSpeed                  ID = "minecraft:flying_speed"
 	TemptRange                   ID = "minecraft:tempt_range"
 	CameraDistance               ID = "minecraft:camera_distance"
+	BelowNameDistance            ID = "minecraft:below_name_distance"
+	NameTagDistance              ID = "minecraft:name_tag_distance"
 	// The 26.3 movement-physics trio a sulfur cube's swallowed block tunes.
 	Bounciness       ID = "minecraft:bounciness"
 	FrictionModifier ID = "minecraft:friction_modifier"